ResearchCalc: Move SleepScore.tsx from ui/ to ui/ResearchCalc/
